my commit file

+++ 0x00-hello_world/6-size.c
@@ -3,7 +3,7 @@
 /**
  * main - Entry point
  *
- * Return : Always 0 (Success)
+ * Return: Always 0 (Success)
  */
 int main(void)
 {
@@ -14,9 +14,9 @@ int main(void)
 	float e;
 
 	printf("Size of char: %lu byte(s)\n",(unsigned long)sizeof(a));
-	printf("Size of int: %lu byte(s)\n", sizeof(b));
-	printf("Size of long int: %lu byte(s)\n", sizeof(c));
-	printf("Size of long long int: %lu byte(s)\n", sizeof(d));
-	printf("Size of float: %lu byte(s)\n", sizeof(e));
+	printf("Size of int: %lu byte(s)\n",(unsigned long) sizeof(b));
+	printf("Size of long int: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("Size of long long int: %lu byte(s)\n", (unsigned long)sizeof(d));
+	printf("Size of float: %lu byte(s)\n", (unsigned long)sizeof(e));
 	return (0);
 }
